Changes applied to action Satya's review comments
