+++ src/SFA.Tl.Matching.Automation.Tests/App.config
@@ -5,17 +5,17 @@
 	</configSections>
 	<appSettings>
 		<add key="Browser" value="chrome" />
-		<add key="BaseUrl" value="XXX" />
-		<add key="DevBaseUrl" value="XXX" />
-		<add key="AdminUserName" value="XXX" />
-		<add key="AdminUserPass" value="XXX" />
-		<add key="StandardUserName" value="XXX" />
-		<add key="StandardUserPass" value="XXX" />
-		<add key="DualUserName" value="XXX" />
-		<add key="DualUserPass" value="XXX" />
-		<add key="NonAuthorisedUserName" value="XXX" />
-		<add key="NonAuthorisedUserPass" value="XXX" />
-		<add key="MathcingServiceConnectionString" value="XXX" />
+		<add key="BaseUrl" value="https://test.industryplacementmatching.education.gov.uk/" />
+		<add key="DevBaseUrl" value="https://tl-at-mtchui-as.azurewebsites.net/" />
+		<add key="AdminUserName" value="Tmatching2" />
+		<add key="AdminUserPass" value="Windmill1" />
+		<add key="StandardUserName" value="Tmatching1" />
+		<add key="StandardUserPass" value="Windmill1" />
+		<add key="DualUserName" value="Tmatching3" />
+		<add key="DualUserPass" value="Windmill1" />
+		<add key="NonAuthorisedUserName" value="fcso" />
+		<add key="NonAuthorisedUserPass" value="Windmill2" />
+		<add key="MathcingServiceConnectionString" value="Server=tcp:tl-test-shared-sql.database.windows.net,1433;Initial Catalog=tl-test-mtch-db;Persist Security Info=False;User ID=mtch-rw-svc;Password=CH6Ztuv9tnAZkxJrC8kxGa2Zq;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=120;" />
 	</appSettings>
