+++ src/SFA.Tl.Matching.Automation.Tests/Project/Tests/Pages/CheckEmployersDetailsPage.cs
@@ -131,9 +131,9 @@ namespace SFA.Tl.Matching.Automation.Tests.Project.Tests.Pages
                 String expectedPhoneNo = employerdetails[2].ToString();
                 String expectedEmail = employerdetails[3].ToString();
 
-                PageInteractionHelper.AssertText(expectedContactName, actualContactName);
-                PageInteractionHelper.AssertText(expectedEmail, actualEmail);
-                PageInteractionHelper.AssertText(expectedPhoneNo, actualPhoneNumber);
+                PageInteractionHelper.VerifyText(expectedContactName, actualContactName);
+                PageInteractionHelper.VerifyText(expectedEmail, actualEmail);
+                PageInteractionHelper.VerifyText(expectedPhoneNo, actualPhoneNumber);
 
                ScenarioContext.Current["_EmployerContactName"] = actualContactName;
             }
